fix(rust): change duration_ms from u64 to u32 for specta TS export compatibility
specta rejects u64 (BigIntForbidden) because it exceeds JS safe integer range;
u32 is sufficient for millisecond durations (max ~49 days) and maps to TS number.

+++ src-tauri/src/models.rs
@@ -80,7 +80,7 @@ pub enum VerificationStatus { NotRun, Running, Passed, Failed, Skipped, Warning 
 pub struct VerificationCheck {
     pub kind: String,
     pub status: VerificationStatus,
-    pub duration_ms: Option<u64>,
+    pub duration_ms: Option<u32>,
     pub log_excerpt: Option<String>,
 }
